refactor(maintenance): extract store kernels and tick policy
