+++ crates/tracedecay/src/daemon.rs
@@ -6,7 +6,7 @@ use std::sync::Arc;
 #[cfg(test)]
 use std::sync::atomic::AtomicUsize;
 use std::sync::atomic::{AtomicBool, Ordering};
-use std::time::{Instant, SystemTime, UNIX_EPOCH};
+use std::time::Instant;
 
 use serde_json::json;
 #[cfg(unix)]
@@ -262,8 +262,6 @@ pub(crate) mod retained_test_support;
 mod shutdown_coordination;
 mod shutdown_orchestration;
 mod shutdown_watchdog;
-#[cfg(feature = "hotpath")]
-pub use shutdown_watchdog::install_hotpath_shutdown_finalizer;
 pub(crate) use core_admission::*;
 pub use core_client::*;
 pub(crate) use core_doctor::*;
@@ -273,6 +271,8 @@ pub(crate) use core_lifecycle::*;
 pub use core_logging::*;
 pub use core_proxy::*;
 pub(crate) use shutdown_coordination::ShutdownStatus;
+#[cfg(feature = "hotpath")]
+pub use shutdown_watchdog::install_hotpath_shutdown_finalizer;
 mod github_credential_lifecycle;
 mod graph_resolution;
 use graph_resolution::retained_project_server_resolver;
